Escape Java using the velocity engine
It is preferable to escape Java through the template engine over
escaping the variables. This ensures none can be forgotten by
accident.

This requires that glue, tags and format are passed as lists rather
then comma seperate values in the pom. Due to the use of default
values splitting format has not been refactored.

Additionally the templates have been formatted such that the
output is a properly formatted java file. This makes debugging
and trouble shooting easier. The integration tests have been
updated accordingly.

Changed the type of GenerateRunnersMojo.cucumberOutputDir to File
and used the build.directory expression to ensure the default
location is always relative to the build directory. As a result
the full path is injected into the generated file. The
integration tests have been updated accordingly.