Add support for custom cucumber plugins
Using the  plugins element we can now support custom cucumber plugins. This should be backwards compatible.

Plugins can also be provided through `-Dcucumber.options`. To avoid over complicating everything this (for now)
has been explicitly limited to build-in cucumber plugins. This was previously implicitly limited.

OverriddenCucumberOptionsParameters and RuntimeOptions have been refactored to allow the overrides to
happen in GenerateRunnersMojo . This was needed because otherwise the default cucumber output directory
needed to be passed down.

A minor error has been corrected with regards to the html formatter. The html formatter create a folder
rather then a file and as such does not need the html extension. The multiple-format integration tests have
been changed to reflect this.