Support cucumber-junit options. Requires cucumber-junit:2.0.0.
By default steps are included in notifications and descriptions. When interacting with cucumber in
an IDE it is nice to see the step executions. However presenting step executions as tests to
various reporters such as surefire results in strange test counts and weird reports.

By passing the no step notification junit option to cucumber this behaviour can be disabled.