Modulo Counter enhancement
This enhancements addresses: https://github.com/temyers/cucumber-jvm-parallel-plugin/issues/135

It parallelisation of cucumber tests to be configured at the continuous integration pipeline level by enabling a user to add a modulo counter to the name of each test.

An additional ModuloCounter is created which counts modulo n where n is an integer extracted from the configured naming pattern using regex "{c:(\d*}}"

This ModuloCounter is used in the same way as the oneUpCounter, however, because the ModuloCounter is configured by the naming pattern, it could not be passed in as a constructor argument to the ClassNamingSchemeFactory, instead it is created in the constructor of the PatternNamingScheme.